Create new parameter set

Create a new parameter set within an rmaverick project. The new parameter set becomes the active set once created.

new_set(project, name = "(no name)", lambda = 1, admix_on = FALSE,
  alpha = 0.1, estimate_alpha = TRUE)

Arguments

project
an rmaverick project, as produced by the function mavproject()

name
the name of the parameter set

lambda
shape parameter of the dirichlet prior on allele frequencies in each subpopulation

admix_on
whether to allow admixture between subpopulations, in which case each gene copy can be assigned to a different subpopulation

alpha
parameter governing the strength of admixture. Higher values give greater probability of a gene copy being assigned to different subpopulations, while as alpha tends to zero we converge back at the no-admixture model in which all gene copies within an individual are constrained to have originated from the same subpopulation

estimate_alpha
whether the value of alpha should be estimated as part of the MCMC, in which case the value alpha is ignored