Run main MCMC

Run the main rmaverick MCMC. Model parameters are taken from the current active parameter set, and MCMC parameters are passed in as arguments. All output is stored within the project.

run_mcmc(project, K = 3, burnin = 100, samples = 1000, rungs = 10,
  GTI_pow = 2, auto_converge = TRUE, converge_test = ceiling(burnin/10),
  solve_label_switching_on = TRUE, coupling_on = TRUE, cluster = NULL,
  pb_markdown = FALSE, silent = !is.null(cluster))

Arguments

project
an rmaverick project, as produced by the function mavproject()

K
the values of K that the MCMC will explore

burnin
the number of burn-in iterations

samples
the number of sampling iterations

rungs
the number of temperature rungs

GTI_pow
the power used in the generalised thermodynamic integration method. Must be greater than 1.1

auto_converge
whether convergence should be assessed automatically every converge_test iterations, leading to termination of the burn-in phase. If FALSE then the full burnin iterations are used

converge_test
test for convergence every convergence_test iterations if auto_converge is being used

solve_label_switching_on
whether to implement the Stevens' solution to the label-switching problem. If turned off then Q-matrix output will no longer be correct, although evidence estimates will be unaffected.

coupling_on
whether to implement Metropolis-coupling over temperature rungs

cluster
option to pass in a cluster environment (see package "parallel")

pb_markdown
whether to run progress bars in markdown mode, in which case they are updated once at the end to avoid large amounts of output.

silent
whether to suppress all console output